pyproject.toml
[tool.poetry]
name = "architecture-python"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "architecture_python", from = "src"}, { include = "src" }]

[tool.poetry.dependencies]
python = "^3.10"
pytest-cov = "^4.1.0"
fastapi = {extras = ["all"], version = "^0.110.0"}
sqlalchemy = "^2.0.28"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
